style: prefix to nwp

+++ act-p0-housekeep.sh
@@ -6,6 +6,7 @@ TZ=UTC; export TZ
 
 # taken from run-prep0.sh
 : ${base:?} ${reftime:?} ${yesterday:?}
+logger --tag p0-housekeep --id=$$ -p news.info "base=${base} reftime=${reftime} yesterday=${yesterday}"
 
 cd $base
 if test -f stop ; then
@@ -18,25 +19,33 @@ ruby /nwp/bin/syslogscan.rb /var/log/syslog.1 > ${yesterday}/syslogscan.ltsv
 
 bash /nwp/bin/mailjis.sh ${yesterday}/syslogscan.ltsv news -s syslogscan-${yesterday}.ltsv news
 
+set +e
+
 (cd ${yesterday}; tar cf logs-${yesterday}.tar *.ltsv)
 
-weekago=$(ruby -rdate -e 'puts(Date.parse(ARGV.first) - 7)' ${reftime})
+weekago=$(ruby -rdate -e 'puts(Date.parse(ARGV.first) - 9)' ${reftime})
 
-for dir in 2*[0-9]
+for dir in 2???-??-?[0-9]
 do
   if [[ $dir < $weekago ]] ; then
-    logger --tag p0-housekeep --id=$$ -p news.notice "rm -rf $dir"
-    rm -rf $dir
+    tfiles=$(find $dir -perm -1000 -print)
+    if [[ -n $tfiles ]] ; then
+      logger --tag p0-housekeep --id=$$ -p news.err -s -- "sticky files $tfiles"
+    else
+      logger --tag p0-housekeep --id=$$ -p news.notice -- "rm -rf $dir"
+      rm -rf $dir
+    fi
   else
     for tar in $dir/*.tar ; do
-      if [[ ! -f ${tar}.gz ]] ; then
+      if [[ -f ${tar} ]] && [[ ! -f ${tar}.gz ]] ; then
         logger --tag p0-housekeep --id=$$ -p news.notice "gzip -9k ${tar}"
         gzip -9k ${tar}
+        chmod o+t ${tar}.gz
       fi
     done
     if [[ $dir < $yesterday ]] ; then
       for tar in $dir/*.tar ; do
-        if [[ -f ${tar}.gz ]] ; then
+        if [[ -f ${tar} ]] && [[ -f ${tar}.gz ]] ; then
           logger --tag p0-housekeep --id=$$ -p news.notice "rm -f ${tar}"
           rm -f ${tar}
         fi
